scheduler: More efficient updateTask
The existing implementation of updateTask removed the old task (if any)
from the node's entry in the heap, and readded the new version of the
task. This is the simplest way to do it, but it's suboptimal. It results
in a lot of churn for most task changes.

Improve this by only updating the node heap entry if actually necessary.

This changes NodeInfo.removeTask and NodeInfo.addTask to be idempotent,
which has the side effect of fixing a bug. Previously, deleted tasks
could mess up resource accounting by adding to avialable resources a
second time.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>