Recover from unrecongized leader
Add an RPC that can resolve a raft node ID into an address to connect
to.

When a node is asked to send a message to a node ID that's not in the
member list, ask a random member to resolve that ID. Open a temporary
connection to the resolved address to send the message. By reenabling
bidirectional communication with the leader, this should allow the
affected node to recognize the configuration change that added the
current leader as a committed log entry, after which it will establish
a long-lived connection to the manager and no longer need this
workaround.

Fixes #436

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>