Add SilenceUsage flag to swarmd command
Without this flag, any error that the manager encounters produces a
usage message. Many errors are not about flags but are about other
things (permissions, failure to connect), etc, so showing a usage
message is not always appropriate. We should trigger a usage message
explicitly when there's an issue parsing a flag, or a missing required
flag.

Fixes: #541

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>