scheduler: Take port allocations into consideration
The scheduler should consider host ports as part of scheduling
decisions. It needs to take into account both requested host ports in
Spec, and host ports that were dynamically assigned in Status.

This adds a PortFilter to the scheduler pipeline, and keeps track of
host ports used on each node in the NodeInfo struct.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>