Handle multi-updates to objects in allocator test
Adjust allocator test code to handle cases where the objects could be
updated more than once. Adjusted the watch object functions to accept
function arguments which can be tested against until true or timeout
rather than trying to estimate the number of updates that can happen to
the object. Will be useful when multiple allocators are added to the
allocator.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>