Manager acceptance
Add methods to the Control service for fetching, listing, and updating
registered certificates.

Add an AcceptancePolicy message. Right now this is basically a stub.

Change the certificate acceptance reconciliation loop to check the
acceptance policy, and when necessary, only issue certificates when the
desired state is set to "issued".

Add "swarmctl cert" subcommand to list, approve, and reject certificate
requests.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>