scheduler: Make addTask and removeTask noops on empty NodeInfo structs
The nodeInfo function returns an empty nodeInfo struct when the node
does not exist in the heap. Rather than check this every time, the
scheduler just plows ahead anyway and relies on nothing actually
happening if it's working with an empty nodeInfo. This keeps the logic a
lot simpler. But addTask/removeTask need to return early if they are
dealing with an empty nodeInfo; otherwise they might try to access a nil
map.

Also, clean up the functions to remove a few levels of indentation.

Signed-off-by: Aaron Lehmann <aaronl@vitelus.com>