orchestrator: Wait for old task to terminate
Change the restart supervisor to wait for the old task to terminate
before setting the desired state of the new one to RUNNING.

If the previous task already completed, or the node is unreachable, we
go to the RUNNING state immediately after any configured restart delay.
We skip the restart delay for tasks that were on drained nodes.

Make the updater use the RestartSupervisor so it can take advantage of
this behavior.

Add a StopGracePeriod option and wire it up in the agent.

Fixes #488
Fixes #577

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>