agent: define single operation controller function
To better control state transitions and error reporting, we've
refactored the runtime into a single method that that takes a task and
returns the next state. By doing this, we centralize the state update
logic and only make the transitions at the Controller call sites. This
should greatly reduce bugs and issues due to disparate status reporting
logic.

This changeset only includes the function definition and tests for the
flow. A later change will hook up the status reporting flow to the rest
of the agent.

Signed-off-by: Stephen J Day <stephen.day@docker.com>