Advertise IP address associated with default route by default
If the admin does not provide a specific --listen-addr value, use the
system's IP address associated with its default route as the address to
hand out to raft peers and agents. This is not guaranteed to be correct,
but it's a much better default than 0.0.0.0, which will never work
outside the same system.

Note that this does not affect the address we actually bind the listener
to. If the --listen-addr flag is left using the default 0.0.0.0 address,
we'll still bind to 0.0.0.0, but have a useful IP to hand out to peers.

Add a warning when starting a manger without specifying --listen-addr.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>