Move some raft settings to Cluster object
Move SnapshotInterval and LogEntriesForSlowFollowers settings inside the
cluster object.

Also create a KeepOldSnapshots setting which controls how many old
snapshots are kept around.

These raft settings are the easier ones to move into the cluster object
because their effect is local, and they don't need to be coordinated
precisely like some of the other settings. In the future we can build on
this by moving the more complex settings to the Cluster object as well.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>