orchestrator: Set desired state to "shutdown" rather than "dead"
When we want a task to stop running, set its desired state to "shutdown"
instead of "dead" so that the container is still available for analysis.
This makes it much easier to debug things like crash loops or incorrect
arguments to a process.

The number of containers to keep around per instance can be controlled
with TaskHistoryRetentionLimit. When this limit is exceeded, the oldest
tasks will be deleted from raft, and therefore from the assignment set
that the agent receives. When old tasks disappear from the assignment
set, the agent will remove their containers.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>