agent: Don't remove tasks from local task map until they disppear from the assignment list
Previously, we were removing tasks from this map in updateStatus once
the tasks reached "dead". The problem with this is that the tasks can
still be present in the assignment list, and they will look like new
tasks that the agent needs to start running. Thus the agent may start a
task that replaces one it was supposed to kill.

By removing this map entry when the task is removed from the assignment
set, it won't be possible to unintentionally restart the task.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>