api: remove Register method from Dispatcher
To simplify the dispatcher API, the Register method is now removed. This
has a number of advantages, including reducing the total number of RPCs.
We also decouple Node lifecycle from the Session lifecycle. If we need
to include data in response to a session creation, we can now simply
include it in a session message and update it as required.

The change is quite simple from a dispatcher prespective: what used to
be the Register method is just called inside of session. There are
probably some optimizations that need to be explored, but it generally
works exactly as it has before.

Signed-off-by: Stephen J Day <stephen.day@docker.com>