Map between raft IDs and node IDs in raft member list
This is an alternative approach to #652 that keeps the raftID<->nodeID
mapping in raft's member list instead of in the Node object.

As with #652, raft IDs are allocated by the cluster instead of by the
joiner.

This may be cleaner because it doesn't involve wrapping the raft joins
with something that updates the node object, and because it avoids
corner cases where the node object may not exist for a manager. The
downside is that knowledge of node IDs leaks inside raft. I'm not sure
which approach is better.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>