Make Manager part of a Node
This makes Node a structure that represents an agent which may also be a
manager.

In this model, we must update the Node structure when a new manager is
added. To support this, a new dispatcher method called PromoteManager
wraps raft joins. This also involves allocating raft IDs on the leader
of the existing cluster, which is something we need to do anyway for
correctness reasons.

To support standalone "swarmctl manager", a node is created for a new
manager if its ID doesn't already represent a registered agent. This
hack could be removed in the future if we transition to a model where
managers are only created through promotion of existing agents.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>