agent: refactor task management flow
We've split the agent task execution flow into several components to
simplify behavior.

Resolve and Do

To ensure proper state manager, we now employment a Resolve and Do
function. Each does an operation on task and returns the state
afterwards. Resolve uses an executor to get the controller and return
task acceptance. Do handles all of the rest of the opertions, including
Prepare, Start, Wait and Shutdown. By centralizing state, we reduce
problems with disparate state management and decrease the surface area
where bugs can occur.

Do was introduced in an earlier commit but the new version correctly
handles issues pausable and unpausable states, as desired by the
orchestrator.

Worker

Worker is the main engine of task execution. It has been separated from
the agent and is now driven solely by task assignement. A worker ensures
that the proper tasks are running and persists tasks and their state to
disk for resilience to restarts.

TaskManager

A task manager is the worker-side reconciliation loop, instantiated per
task. It coordinates asynchronous operations, propagated changes in
state to a reporter and manages updates. It is structured to minimize
the concurrent operations on an exec.Controller.

Reporter

Reliable reporting is now handled by a separate component. It continues
to retry sending task states until an update is received or the agent is
shutdown.

DB

BoltDB has been introduced as a worker-side persistent state store for
tasks. Each task has its own bucket, with status, task data and
assignement stored in separate keys. Rather than create a storage
object, we use the rich bolt.Tx interface directly with a set of clear
helper functions. Database access is *only* done through the worker.

Signed-off-by: Stephen J Day <stephen.day@docker.com>