agent: define the concept of a taskManager
A task manager controls the and reporting of a single task. In the past,
this has been managed directly in the agent, leading to complex and
fragile code. With this component, the actions surrounding a task are
well-separated from other concerns. We bring back a reporter-style
interface, but it is limited to just the agent package, rather than
being an integration in the exec package.

We also introduce the concept of a "modifier" to manage the
serialization of state mutations. Modification requests are sent over a
channel, with a response channel. The operation is then consumed by an
internal select loop, where it has free access to any state variables
present. Once the operation proceeds, the error is passed back over the
channel.

Signed-off-by: Stephen J Day <stephen.day@docker.com>