Generate authorization wrapper for remotely exposed control API functions
The control API functions need to be exposed on a TCP socket so they can
be routed to the leader within the raft cluster. However, in the TCP
case, they need to check for a valid manager certificate. Support this
by codegenning a wrapper that checks for authorization. Register this
wrapper on the TCP listener instead of the control API server itself.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>