agent/exec/container: correctly pass command arguments
In the ContainerSpec, we've defined the runtime arguments as Command and
Args since this is a more familiar concept outside of the Docker world.
It also provides better behavior in that it is clear when arguments are
passed to an Entrypoint. We do this by ensuring that only Args are
passed to an Entrypoint. If Command is defined, both Entrypoint and Cmd
are replaced on the container to allow one to completely ignore the
entrypoint and run whatever is desired.

Signed-off-by: Stephen J Day <stephen.day@docker.com>