Don't proxy GetRootCACertificate method
Add an extension option "dont_proxy" that lets us mark particular RPCs
as ones that should always be handled locally. In the case of
GetRootCACertificate, this is a special RPC that must allow
unauthenticated clients, and the proxy logic includes authentication
that prevents this from working.

In the future, we can take the concept of method options further, and
have options in the protobuf files indicating what authorization roles
(if any) each RPC requires. Combined with the authentication wrapper in
PR #657, this would avoid the need for error-prone manual authorization
checks at the top of each handler's implementation.

In combination with #644 and #659, this commit allows an agent to join
any manager, not just the original root CA. It almost allows a manager
to join any manager, but this is currently blocked by lack of OU
forwarding in the raft proxy (see
https://github.com/docker/swarm-v2/pull/644/files#r64508984).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>