Code generated RPC handler authorization
Extend the grpc method options with a new TLSAuthorization option. This
allows us to specify which TLS roles are allowed for each method inside
the protobufs instead of needing to remember to insert a check in each
handler.

Change authenticatedwrapper to honor this setting in the code it
generates.

Wire up authenticatedwrapper in front of each service that's exposed
over TCP. Add unit test coverage making sure these APIs are not
accessible without a certificate.

Change raftproxy to forward both the CN and the OUs.
authenticatedwrapper will accept either a direct request from an allowed
OU, or a forwarded request from an allowed OU. raftproxy no longer does
any authorization, leaving that to the remote side. This will allow CA
methods to work correctly with raftproxy.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>