Make raft and the certificate share the same node ID
Before this PR, there were two separate ID spaces for raft and
certificate CNs. Certificates used a base-36 ID with 128 bits of
entropy. Raft used 64 bit integers.

This PR changes certificates to use 64-bit identifiers in the CNs, so
those IDs can be shared by raft. The IDs are still expressed in base-36
in the certificates and in the API.

Raft no longer generates IDs - now they are taken from the certs.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>