Merge certificates into Node object
To support a model where certificates that haven't been approved yet
appear as pending nodes, remove the concept of RegisteredCertificate and
instead put a Certificate message on the Node object.

This splits CA RPCs that have become node-specific into a NodeCA service.
In the future, we could add certs to other object types, and create
corresponding VolumeCAServer, NetworkCAServer, etc. services.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
Signed-off-by: Diogo Monica <diogo.monica@gmail.com>