Revert "Ensuring that non-accepted nodes don't get schedulled tasks"
This reverts commit 6d32959095d2cbfb82f1255f51b8b945571bd02d.

The breaks autoacceptance, since the spec doesn't get updated when we
autoaccept a node.

The scheduler shouldn't be aware of TLS certificate issuance. It should
just check Status.State as before. When we block a node (currently not
implemented), we should disconnect it and prevent it from reregistering.
Its status will be something other than ACTIVE, so the scheduler will do
the right hting.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>