Add constrained scheduler benchmarks
This benchmarks scheduling with constraints, where only 1/3 of the nodes
are eligible to run the tasks.

Under the current implementation, this converges towards the worst-case
linear behavior, because the nodes that meet the constraints end up at
the end of the heap, and findMin has to scan much of the heap to find
them.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>