[ca] Fix GetLocalRootCA when key not found
According to the function comments, GetLocalRootCA should, when only a CA Cert
exists locally without the corresponding private key, return a RootCA without
a signer. Instead, it was returning an error that the private key was invalid.
This patch fixes this issue to match the behavior described in the comments.

Docker-DCO-1.1-Signed-off-by: Josh Hawn <josh.hawn@docker.com> (github: jlhawn)