store: Add an "OR" combinator
This allows queries like:

store.FindTasks(store.Or(store.ByName(name), store.ByNodeID(nodeID), store.ByServiceID(serviceID)))

I also considered adding AND, but I don't think we should support it at
this time. The best way to compute the intersection is to start with the
most specific dimension and then narrow the results from there, but a
generic AND operator would only be able to do the query in a naive way.
It's probably better for callers to implement their own AND logic.
But we might add AND in the future.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>