scheduler: Refactor Pipeline and filters to allow per-task precomputation
The current approach creates a new Pipeline for each task we want to
assign. Then we iterate over available nodes and for each filter,
`Check(t *api.Task, n *NodeInfo)` is called. If there is any information
derived from the task, it must be recomputed on every call to Check.

This commit changes the scheduler to use a single Pipeline. Pipeline and
the filters now have SetTask methods to start processing a new task, and
do any necessary precomputation. The Check function only takes a
NodeInfo struct.

This approach avoids redundant computation on each call to Check because
task-specific information can be stored inside the filter objects, and
replaced on the next call to SetTask. It also avoids the allocs involved
in creating a new Pipeline for each task.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>