Update go-events dependency
This vendors a new go-events, which changes the warning log line when a
queue is closed with outstanding events to a debug log line. This
suppresses some spurious log output when shutting down a manager, and
also from unit tests and benchmarks that don't fully read from their
event queues.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>