allocator: Batch task updates
Currently, the allocators use a separate transaction for each task
update in some cases. Each one of these entails a raft round trip. So if
the orchestrator creates 1000 tasks for a new service, it could take a
long time to move all of them to the allocated state.

This commit changes it to batch these task updates. Instead of updating
a task within its own transaction, changes are always made as part of
procUnallocatedTasksNetwork / procUnallocatedTasksVolume, which are now
triggered on a transaction commit.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>