raft: Attempt to fix flaky tests
I've seen raft tests fail occasionally, particularly TestRaftSnapshot
and TestRaftSnapshotRestart. These usually seem to fail becuase the last
proposal does not appear in the store on one of the nodes, even after
several seconds of polling.

I don't know what's going wrong, and it's very hard to track down
because I can't reproduce the failures at will. However, my best guess
is that something requires time to pass to behave properly, and because
of our mock clock source, this might not be happening. For example, a
send may need to be retried in some rare cases, but this would not
happen until the next tick.

This commit attempts to make time pass for the raft state machine during
the polling. I think the likelyhood that this fixes the problems is
relatively low, but it's worth a try.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>