Remove Manager controlapi handlers
Move everything under Node*. Listing nodes or retrieving a node will
show the manager information if this node is a manager.

NOTE: This commit removes the ability to remove managers from the raft
cluster. This should be reintroduced through demotion in a followup PR.
A future NodeRemove method should remove non-manager nodes, but insist
that a manager be demoted before allowing its removal.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>