api: simplify volumes support
Previously, Volumes were a top-level object in the Swarm cluster. After
analyzing several uses, it turned out most of the complexity of
distributed volumes is completely deferred to volume plugins. Supporting
Swarm managed volumes would require too many changes to the way volumes
work to be supportable at this point.

To deal with this, volumes calls are now just a `VolumeTemplate`, which
was added in a previous PR. This changeset goes further and removes the
entire top-level Volume object to avoid confusion.

It is likely we will add this in the future, but having a
half-implementation without further analysis is unwise.

Signed-off-by: Stephen J Day <stephen.day@docker.com>