api: remove disconnect field from session message
After some close study of session handling and GRPC, we found it to be
sufficient to close the Server's transport stream to force a
reconnection at the agent-level. Under this realization, we have removed
the `disconnect` field from `api.SessionMessage`.

Connection rebalancing is not yet implemented, but we have confirmed
that the current behavior is sufficient to support in the future without
changes in the agent. The detected connection closer on the client-side
forces GRPC to call into the picker and select a new manager node.

Signed-off-by: Stephen J Day <stephen.day@docker.com>