Assign separate raft ID at join time
Reverse the decision to merge the node ID and raft ID concepts. Instead,
have the leader assign a random raft ID at join time. This allows a node
to be demoted and then be promoted again.

This splits the Raft service into Raft and RaftMembership, so
RaftMembership can use the raft proxy to make sure Join/Leave always run
on the leader. This was necessary anyway, but this change makes it
especially necessary.

Change Node.Manager to Node.ManagerStatus, which is now a flattened
version of RaftMember.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>