orchestrator: Fix service scaling deadlock
The updater's Update method was being called from inside store.Batch,
which holds the store write lock. In some cases this would call Cancel,
which blocks until goroutines from the previous update finish, and this
can deadlock since those goroutines may be blocking on acquiring the
store write lock. The solution is to call Update outside of the
store.Batch callback, since there's no reason for this to be called
inside the batch.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>