agent/exec: revise error handling approach
Errors returned from controller methods are now all treated as fatal by
the the task controller. The defers most retries to the orchestration
layer, where they can be better handled. We also cement the concept of a
"Temporary" error, allowing an Executor/Controller to signal to the
caller that the error is actually temporary. We can mark these as
temporary as we find cases where it makes more sense.

Signed-off-by: Stephen J Day <stephen.day@docker.com>