raft: Fix edge case panic
The logic in sendToMember is slightly wrong. We look for a member other
than ourself to query, but there's no error case if we are the only
member in the list. Add a check to make sure

This code should normally not be reached if we are the only member, but
appears to run when an invalid address is passed to Join. A separate
change will prevent nodes with invalid addresses from being added to the
cluster.

Related to #983

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>