manager: Return an error if no TCP listen address is provided
One of the necessary parameters for creating a manager is the TCP
address where it will listen for GRPC traffic. It's okay to use a
wildcard address like 0.0.0.0:4242, but the port is still important, so
an empty string isn't a usable value.

Return an error from (*Manager).New if no listen address (or listener)
is provided. This avoids bad behavior in raft if the address is omitted.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>