raft: Avoid deadlock on demotion
Demotion depends on raft's Run function noticing that it has been
removed from the cluster, and returning an error.

Before it returns the error, it calls stop() to shut down raft. This
could deadlock in some cases if a proposal was in progress.

Change the context passed to Propose so the proposal will get cancelled
when stop() cancels the context.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>