ca: Verify that CSR in NodeCertificateStatus response matches
A race condition is possible when we renew a certificate at the same
time as another update is being made to a node object.
NodeCertificateStatus is meant to wait until the certificate is issued,
but it does so by waiting for an update to the node object that has the
certificate issuance status set to "issued". If another certificate is
issued shortly before the renewal request, or an unrelated change to the
node object happens at the same time, NodeCertificateStatus may return
the old certificate instead of waiting for the new one.

To work around this, the caller of NodeCertificateStatus must verify
that it's getting a certificate corresponding to the CSR it submitted,
and retry if the CSR does not match.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>