ca: Retry certificate retrieval operations
GetRemoteCA and GetRemoteSignedCertificate use GRPC Pickers, which do a
pretty good job of trying different remote hosts until they find one
that accepts the connection. However, there is a case they don't cover.
If the first manager they contact was recently demoted and is in the
process of shutting down, it will accept GRPC connections, but its CA
endpoints will return errors.

To work around this corner case, add bounded retry loops around the call
sites of GetRemoteCA and GetRemoteSignedCertificates. If a call fails
with an error, fetching the cert will be retried up to 4 more times
before giving up.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>