orchestrator: Trigger rolling updates when endpoints change
This checks the task's Endpoint field against the service's Endpoint
field to see if tasks need to be replaced to take into account new
endpoint settings.

Endpoint is not a spec, but it is copied directly from the service, so
it's safe to do this comparison. It's actually better than comparing the
specs, because if an arbitrary port is allocated at runtime, this will
ensure that the task and service share the same port value.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>