raft: Avoid connecting to cluster members that have been removed
Add a check to avoid opening a connection to a cluster member that has
been removed. This avoids ugly log messages about failed connections
when using --force-new-cluster.

Propagate the removed list on Join as well, so that newly joining
members won't connect to removed members as the process the
configuration changes that originally added them.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>