[ca] Enable node cert signing by external CA
A cluster can now be configured to make node certificate signing requests to
one or more external CA servers which implement the CFSSL sign API. The first
manager in the cluster must be initialized with a root certificate and node
key and certificate issued by the external CA. The cluster store can also be
updated to modify the URLs of the external CA API endpoints.

This patch also updates the tests in the ca package with a custom TestMain()
function which runs all tests twice: once with a local signer and another run
which is configured to us an external signing server which is setup by the
testutils.

Docker-DCO-1.1-Signed-off-by: Josh Hawn <josh.hawn@docker.com> (github: jlhawn)