Follow logs from service
One feature that at the moment we don't have but could be good to have
is a command to follow logs for a particular service.
With this PR I am trying to implement this feature and also grab some
information about this feature.

This is my first contribution and plese let me know if this idea is
completly wrong.

If someone has feedback for my this is a checklist about my next steps

- [x] Bootstrap new command `swarmctl service logs <service-name>`
- [x] Grab containerId for each task running
- [ ] obtain logs from this list of containers
- [ ] Print something of good to see
- [ ] Implement flags like `-f` to follow or not a stream

Signed-off-by: Gianluca Arbezzano gianarb92@gmail.com
Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>