orchestrator: Better balancing on scale down
The existing replicated orchestrator scales down by deleting tasks with
the highest slot numbers. This means that if you have a service with the
following tasks:

    task slot 1 on node A
    task slot 2 on node A
    task slot 3 on node A
    task slot 4 on node B
    task slot 5 on node C

... then scaling from 5 replicas to 3 replicas will shut down all tasks
on nodes B and C, and you end up with 3 tasks on node A.

Improve downscaling to be aware of how many tasks are on each node, and
prefer to shut down tasks in a way that balances the resulting tasks.
After this change, the result in the example above would be one task
each on nodes A, B, and C.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>