orchestrator: Fix rolling update race that could create extra tasks
The updateTask method atomically shuts down a task and replaces it with
a new one. Calls to updateTask are queued through a channel, and this
happens outside a channel. It's possible that once updateTask is called,
the old task has already been shut down. In this case, shutting down the
task is a no-op, but creating a task adds a new task that shouldn't
exist, and puts us in a bad state. We need to check that the task hasn't
already been shut down to avoid this problem.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>