Move swarmctl, swarmd, and swarm-bench to examples
swarmctl and swarmd are not the recommended way to run swarmkit, and it
has been requested several times that they be moved to an 'examples'
directory so users are less likely to try them. The cmd directory has
been renamed to 'tools' to better reflect its purpose
The Makefile has been altered accordingly. A README has been added to
examples to explain this change

Signed-off-by: Drew Erny <drew.erny@docker.com>