Keep a copy of EndpointSpec in Task for update
To trigger rolling update reliably and more importantly to not trigger
false rolling updates, updater should compare against change in
EndpointSpec between service and what is present in task. For that task
needs to have a copy of EndpointSpec. Made the necessary changes to fix
rolling update of a service.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>