Use netlink(on linux) to determine listen address
Also add fallback for v6-only networks.

The current method opens a connection to a public server and grabs the
local address. This won't work on private only clusters and is just is
not great to call out to the internet.
Change this to use netlink when on Linux. The method used by the netlink
pkg to determine the route is similar to `ip route get` and does not
require calling out to the internet.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>