api: clarify Command/Args distinction
To address some issues with using `Command`/`Args`, the field
documentation has been updated to clarify that Command should only be
used when overriding Entrypoint. For most docker uses, only `Args`
should ever be used, as is done in `runc`.

`Args` is now appended to `Command` if both are present.

A `--command` flag has been added to be able to test that the feature is
working correctly across systems.

Signed-off-by: Stephen J Day <stephen.day@docker.com>