orchestrator: Prefer to remove non-running tasks when scaling down
We already make some attempts in the orchestrator to prefer certain
tasks over others when scaling down. The current logic is focused on
maintaining balance across the various nodes when scaling down - i.e.
not killing all the tasks on one node and leaving several running on
another node.

Another useful heuristic is to prefer removing non-running tasks over
running tasks. If our number of replicas is four, and we have two
starting tasks and two running tasks, scaling down to two replicas
should prefer to preserve the running tasks instead of causing a
temporary service interruption, all other things being equal.

This change prioritizes preservation of running tasks (vs
starting/completed/failed tasks) after node balance. So if the
orchestartor has to break a tie between nodes which could give up a
task during scale down, it will choose tasks which are not running over
tasks that are running.

It makes sense to make this a secondary factor behind node balance,
because removing a running task is not a big deal in the long term - the
state will converge on all tasks running. But if we were to sacrifice
balance between nodes, there's currently no active rebalancing to
restore this.

Fixes #1095

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>