raft membership: Fix promote/demote workflow
This fixes the update node workflow when the Manager is being
demoted by making sure that the role is only updated when the
node has been removed successfully.

This also fixes the problem of scaling down to one manager by
fixing the check on 'CanRemoveMember' which was incorrectly
accounting for the number of reachable nodes amongst the quorum.

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>