api: use ReadOnly rather than Writable
Overwelmingly, the feedback has been that mounts should be writable by
default, rather than readonly. This change is being made across
engine-api, docker and swarmkit to facilitate this preference.

This has some implications for upgrade between versions, in that
services that were running before an upgrade will now have their
writable mounts be read only. The suggested fix is to destroy the
existing service and replace with a new service.

If this change does not make it to the GA, DO NOT MERGE, close this PR
and forget we ever had this idea.

Signed-off-by: Stephen J Day <stephen.day@docker.com>