scheduler: Fix panic processing preassigned tasks on node removal
processPreassignedTasks runs through a list of preassigned global tasks
and tries to see if they can be advanced to the next state. If a node is
deleted, there is a brief window where the global tasks assigned to that
node have not been deleted yet. processPreassignedTasks may call
taskFitNode on one of these global tasks, and since taskFitNode does not
check if the nodeInfo call returns a valid node from the heap, this may
cause a nil pointer dereference.

Change nodeInfo to return an error, and only use the nodeInfo struct it
returns if it returns a nil error.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>