Remove dial-based address detection logic
This didn't prove as usable as we hoped, since often cloud setups have
separate network interfaces for internal and external communication.
Instead, we're taking the approach of requiring the user to specify an
address or interface if there's any ambiguity.

Remove the logic that finds the IP address associated with the default
route, since it will no longer be used.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>